requirements.txt
Django==5.0.7
psycopg2-binary==2.9.9
Pillow==10.4.0
python-decouple==3.8
whitenoise==6.6.0
django-ckeditor==6.7.0
django-imagekit==5.0.0
gunicorn==21.2.0
dj-database-url==2.1.0
